requirements.txt
fastapi
uvicorn
pydantic
requests
google-genai
rich
termcolor
openai
anthropic
dotenv
setuptools